@@ -1,7 +1,7 @@
 {
   "name": "gatsby-plugin-netlify-cms",
   "description": "A Gatsby plugin which generates the Netlify CMS single page app",
-  "version": "5.9.0-next.0",
+  "version": "5.10.0-next.0",
   "author": "Shawn Erquhart <shawn@erquh.art>",
   "bugs": {
     "url": "https://github.com/gatsbyjs/gatsby/issues"
@@ -15,12 +15,12 @@
     "lodash": "^4.17.21",
     "mini-css-extract-plugin": "1.6.0",
     "netlify-identity-widget": "^1.9.1",
-    "webpack": "^5.23.0"
+    "webpack": "^5.35.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.14.0",
     "@babel/core": "^7.14.0",
-    "babel-preset-gatsby-package": "^1.9.0-next.0",
+    "babel-preset-gatsby-package": "^1.10.0-next.0",
     "cross-env": "^7.0.3",
     "react": "^16.12.0",
     "react-dom": "^16.12.0"
