@@ -3,7 +3,7 @@ import { graphql } from "gatsby"
 
 export default ({ data }) => {
   if (!data?.allTest?.nodes) {
-    throw new Error("Wrong data")
+    throw new Error("Wrong data: " + JSON.stringify(data))
   }
   return <div>{JSON.stringify(data)}</div>
 }
@@ -13,7 +13,7 @@ export const query = graphql`
     allTest(
       filter: { nodeNum: { gt: $pageNum, lt: $pagesTotal } }
       sort: $sort
-      limit: 5
+      limit: 100
     ) {
       nodes {
         nodeNum
