@@ -3,14 +3,14 @@ import { graphql } from "gatsby"
 
 export default ({ data }) => {
   if (!data?.allTest?.nodes) {
-    throw new Error("Wrong data")
+    throw new Error("Wrong data: " + JSON.stringify(data))
   }
   return <div>{JSON.stringify(data)}</div>
 }
 
 export const query = graphql`
   query($pagesLeft: Int, $sort: TestSortInput) {
-    allTest(filter: { nodeNum: { gt: $pagesLeft } }, sort: $sort, limit: 5) {
+    allTest(filter: { nodeNum: { gt: $pagesLeft } }, sort: $sort, limit: 100) {
       nodes {
         nodeNum
         text
