@@ -3,7 +3,7 @@ import { graphql } from "gatsby"
 
 export default ({ data }) => {
   if (!data?.allTest?.nodes) {
-    throw new Error("Wrong data")
+    throw new Error("Wrong data: " + JSON.stringify(data))
   }
   return <div>{JSON.stringify(data)}</div>
 }
@@ -11,9 +11,9 @@ export default ({ data }) => {
 export const query = graphql`
   query($pageNum: Int, $sort: TestSortInput) {
     allTest(
-      filter: { nodeNumReversed: { lt: $pageNum } }
+      filter: { nodeNum: { lt: $pageNum } }
       sort: $sort
-      limit: 5
+      limit: 100
     ) {
       nodes {
         nodeNum
