@@ -3,14 +3,18 @@ import { graphql } from "gatsby"
 
 export default ({ data }) => {
   if (!data?.allTest?.nodes) {
-    throw new Error("Invalid data")
+    throw new Error("Wrong data: " + JSON.stringify(data))
   }
   return <div>{JSON.stringify(data)}</div>
 }
 
 export const query = graphql`
-  query($fooBarValues: [String!], $sort: TestSortInput) {
-    allTest(filter: { testIn: { nin: $fooBarValues } }, sort: $sort, limit: 5) {
+  query($fooBarArray: [String!], $sort: TestSortInput) {
+    allTest(
+      filter: { fooBar: { nin: $fooBarArray } }
+      sort: $sort
+      limit: 100
+    ) {
       nodes {
         nodeNum
         text
