@@ -26,20 +26,21 @@ function mergePageEntry(cachedPage, newPageData) {
 class DevLoader extends BaseLoader {
   constructor(asyncRequires, matchPaths) {
     const loadComponent = chunkName => {
-      if (!asyncRequires.components[chunkName]) {
+      if (!this.asyncRequires.components[chunkName]) {
         throw new Error(
           `We couldn't find the correct component chunk with the name "${chunkName}"`
         )
       }
 
       return (
-        asyncRequires.components[chunkName]()
+        this.asyncRequires.components[chunkName]()
           .then(preferDefault)
           // loader will handle the case when component is error
           .catch(err => err)
       )
     }
     super(loadComponent, matchPaths)
+    this.asyncRequires = asyncRequires
 
     const socket = getSocket()
 
@@ -58,6 +59,10 @@ class DevLoader extends BaseLoader {
     } else if (process.env.NODE_ENV !== `test`) {
       console.warn(`Could not get web socket`)
     }
+  }
+
+  updateAsyncRequires(asyncRequires) {
+    this.asyncRequires = asyncRequires
   }
 
   loadPage(pagePath) {
