@@ -1,22 +1,21 @@
 import { IGatsbyNode } from "../../redux/types"
-import { GatsbyGraphQLType } from "../../.."
-import { prepareRegex } from "../../utils/prepare-regex"
-import { makeRe } from "micromatch"
 import { getValueAt } from "../../utils/get-value-at"
 import _ from "lodash"
 import {
   DbQuery,
   IDbQueryQuery,
   IDbQueryElemMatch,
+  IInputQuery,
+  FilterValueNullable,
   objectToDottedField,
   createDbQueriesFromObject,
   prefixResolvedFields,
+  prepareQueryArgs,
 } from "../common/query"
 import {
   FilterOp,
   FilterCacheKey,
   FiltersCache,
-  FilterValueNullable,
   ensureEmptyFilterCache,
   ensureIndexByQuery,
   ensureIndexByElemMatch,
@@ -25,31 +24,11 @@ import {
   IFilterCache,
 } from "./indexing"
 import { IGraphQLRunnerStats } from "../../query/types"
+import { IRunQueryArgs, IQueryResult } from "../types"
+import { GatsbyIterable } from "../common/iterable"
 
-// The value is an object with arbitrary keys that are either filter values or,
-// recursively, an object with the same struct. Ie. `{a: {a: {a: 2}}}`
-interface IInputQuery {
-  [key: string]: FilterValueNullable | IInputQuery
-}
-// Similar to IInputQuery except the comparator leaf nodes will have their
-// key prefixed with `$` and their value, in some cases, normalized.
-interface IPreparedQueryArg {
-  [key: string]: FilterValueNullable | IPreparedQueryArg
-}
-
-interface IRunFilterArg {
-  gqlType: GatsbyGraphQLType
-  queryArgs: {
-    filter: Array<IInputQuery> | undefined
-    sort:
-      | { fields: Array<string>; order: Array<boolean | "asc" | "desc"> }
-      | undefined
-  }
-  firstOnly: boolean
-  resolvedFields: Record<string, any>
-  nodeTypeNames: Array<string>
+export interface IRunFilterArg extends IRunQueryArgs {
   filtersCache: FiltersCache
-  stats: IGraphQLRunnerStats
 }
 
 /**
@@ -81,37 +60,6 @@ function createFilterCacheKey(
 
   // Note: the separators (`,` and `/`) are arbitrary but must be different
   return typeNames.join(`,`) + `/` + paths.join(`,`) + `/` + comparator
-}
-
-function prepareQueryArgs(
-  filterFields: Array<IInputQuery> | IInputQuery = {}
-): IPreparedQueryArg {
-  const filters = {}
-  Object.keys(filterFields).forEach(key => {
-    const value = filterFields[key]
-    if (_.isPlainObject(value)) {
-      filters[key === `elemMatch` ? `$elemMatch` : key] = prepareQueryArgs(
-        value as IInputQuery
-      )
-    } else {
-      switch (key) {
-        case `regex`:
-          if (typeof value !== `string`) {
-            throw new Error(
-              `The $regex comparator is expecting the regex as a string, not an actual regex or anything else`
-            )
-          }
-          filters[`$regex`] = prepareRegex(value)
-          break
-        case `glob`:
-          filters[`$regex`] = makeRe(value)
-          break
-        default:
-          filters[`$${key}`] = value
-      }
-    }
-  })
-  return filters
 }
 
 /**
@@ -315,23 +263,17 @@ function collectBucketForElemMatch(
  * Filters and sorts a list of nodes using mongodb-like syntax.
  *
  * @param args raw graphql query filter/sort as an object
- * @property {boolean} args.firstOnly true if you want to return only the first
- *   result found. This will return a collection of size 1. Not a single element
- * @property {{filter?: Object, sort?: Object} | undefined} args.queryArgs
+ * @property {{filter?: Object, sort?: Object, skip?: number, limit?: number} | undefined} args.queryArgs
  * @property {FiltersCache} args.filtersCache A cache of indexes where you can
  *   look up Nodes grouped by a FilterCacheKey, which yields a Map which holds
  *   an arr of Nodes for the value that the filter is trying to query against.
  *   This object lives in query/query-runner.js and is passed down runQuery.
- * @returns Collection of results. Collection will be limited to 1
- *   if `firstOnly` is true
+ * @returns Collection of results. Collection will be sliced by `skip` and `limit`
  */
-export function runFastFiltersAndSort(
-  args: IRunFilterArg
-): Array<IGatsbyNode> | null {
+export function runFastFiltersAndSort(args: IRunFilterArg): IQueryResult {
   const {
-    queryArgs: { filter, sort } = {},
+    queryArgs: { filter, sort, limit, skip = 0 } = {},
     resolvedFields = {},
-    firstOnly = false,
     nodeTypeNames,
     filtersCache,
     stats,
@@ -339,27 +281,33 @@ export function runFastFiltersAndSort(
 
   const result = convertAndApplyFastFilters(
     filter,
-    firstOnly,
     nodeTypeNames,
     filtersCache,
     resolvedFields,
     stats
   )
 
-  return sortNodes(result, sort, resolvedFields, stats)
+  const sortedResult = sortNodes(result, sort, resolvedFields, stats)
+  const totalCount = async (): Promise<number> => sortedResult.length
+
+  const entries =
+    skip || limit
+      ? sortedResult.slice(skip, limit ? skip + (limit ?? 0) : undefined)
+      : sortedResult
+
+  return { entries: new GatsbyIterable(entries), totalCount }
 }
 
 /**
- * Return a collection of results. Collection will be limited to 1 if `firstOnly` is true
+ * Return a collection of results.
  */
 function convertAndApplyFastFilters(
-  filterFields: Array<IInputQuery> | undefined,
-  firstOnly: boolean,
+  filterFields: IInputQuery | undefined,
   nodeTypeNames: Array<string>,
   filtersCache: FiltersCache,
   resolvedFields: Record<string, any>,
   stats: IGraphQLRunnerStats
-): Array<IGatsbyNode> | null {
+): Array<IGatsbyNode> {
   const filters = filterFields
     ? prefixResolvedFields(
         createDbQueriesFromObject(prepareQueryArgs(filterFields)),
@@ -393,11 +341,7 @@ function convertAndApplyFastFilters(
     // If there is no filter then the ensureCache step will populate this:
     const cache = filterCache.meta.orderedByCounter as Array<IGatsbyNode>
 
-    if (firstOnly || cache.length) {
-      return cache.slice(0)
-    }
-
-    return null
+    return cache.slice(0)
   }
 
   const result = applyFastFilters(filters, nodeTypeNames, filtersCache)
@@ -405,9 +349,6 @@ function convertAndApplyFastFilters(
   if (result) {
     if (stats) {
       stats.totalIndexHits++
-    }
-    if (firstOnly) {
-      return result.slice(0, 1)
     }
     return result
   }
@@ -447,14 +388,17 @@ function filterToStats(
  * Returns same reference as input, sorted inline
  */
 function sortNodes(
-  nodes: Array<IGatsbyNode> | null,
+  nodes: Array<IGatsbyNode>,
   sort:
-    | { fields: Array<string>; order: Array<boolean | "asc" | "desc"> }
+    | {
+        fields: Array<string>
+        order: Array<boolean | "asc" | "desc" | "ASC" | "DESC">
+      }
     | undefined,
   resolvedFields: any,
   stats: IGraphQLRunnerStats
-): Array<IGatsbyNode> | null {
-  if (!sort || !nodes || nodes.length === 0) {
+): Array<IGatsbyNode> {
+  if (!sort || sort.fields?.length === 0 || !nodes || nodes.length === 0) {
     return nodes
   }
 
