@@ -3,12 +3,20 @@ import { RootDatabase, open } from "lmdb-store"
 import { ActionsUnion, IGatsbyNode } from "../../redux/types"
 import { updateNodes } from "./updates/nodes"
 import { updateNodesByType } from "./updates/nodes-by-type"
-import { IDataStore, IGatsbyIterable, ILmdbDatabases } from "../types"
+import { IDataStore, ILmdbDatabases, IQueryResult } from "../types"
 import { emitter, replaceReducer } from "../../redux"
+import { GatsbyIterable } from "../common/iterable"
+import {
+  IRunFilterArg,
+  runFastFiltersAndSort,
+} from "../in-memory/run-fast-filters"
 
 const rootDbFile =
   process.env.NODE_ENV === `test`
     ? `test-datastore-${
+        // FORCE_TEST_DATABASE_ID will be set if this gets executed in worker context
+        // when running jest tests. JEST_WORKER_ID will be set when this gets executed directly
+        // in test context (jest will use jest-worker internally).
         process.env.FORCE_TEST_DATABASE_ID ?? process.env.JEST_WORKER_ID
       }`
     : `datastore`
@@ -73,21 +81,25 @@ function getNodesByType(type: string): Array<IGatsbyNode> {
   return result ?? []
 }
 
-function iterateNodes(): IGatsbyIterable<IGatsbyNode> {
+function iterateNodes(): GatsbyIterable<IGatsbyNode> {
   // Additionally fetching items by id to leverage lmdb-store cache
   const nodesDb = getDatabases().nodes
-  return nodesDb
-    .getKeys({ snapshot: false })
-    .map(nodeId => getNode(nodeId)!)
-    .filter(Boolean)
+  return new GatsbyIterable(
+    nodesDb
+      .getKeys({ snapshot: false })
+      .map(nodeId => getNode(nodeId)!)
+      .filter(Boolean)
+  )
 }
 
-function iterateNodesByType(type: string): IGatsbyIterable<IGatsbyNode> {
+function iterateNodesByType(type: string): GatsbyIterable<IGatsbyNode> {
   const nodesByType = getDatabases().nodesByType
-  return nodesByType
-    .getValues(type)
-    .map(nodeId => getNode(nodeId)!)
-    .filter(Boolean)
+  return new GatsbyIterable(
+    nodesByType
+      .getValues(type)
+      .map(nodeId => getNode(nodeId)!)
+      .filter(Boolean)
+  )
 }
 
 function getNode(id: string): IGatsbyNode | undefined {
@@ -113,6 +125,10 @@ function countNodes(typeName?: string): number {
     count++
   })
   return count
+}
+
+async function runQuery(args: IRunFilterArg): Promise<IQueryResult> {
+  return Promise.resolve(runFastFiltersAndSort(args))
 }
 
 let lastOperationPromise: Promise<any> = Promise.resolve()
@@ -157,6 +173,7 @@ export function setupLmdbStore(): IDataStore {
     iterateNodesByType,
     updateDataStore,
     ready,
+    runQuery,
 
     // deprecated:
     getNodes,
