@@ -1,5 +1,9 @@
 import { Database } from "lmdb-store"
 import { IGatsbyNode } from "../redux/types"
+import { GatsbyGraphQLType } from "../../index"
+import { IInputQuery } from "./common/query"
+import { IGraphQLRunnerStats } from "../query/types"
+import { GatsbyIterable } from "./common/iterable"
 
 export type NodeId = string
 export type NodeType = string
@@ -9,13 +13,28 @@ export interface ILmdbDatabases {
   nodesByType: Database<NodeId, NodeType>
 }
 
-// Note: this type is compatible with lmdb-store ArrayLikeIterable
-export interface IGatsbyIterable<T> extends Iterable<T> {
-  [Symbol.iterator](): Iterator<T>
-  map<U>(fn: (entry: T) => U): IGatsbyIterable<U>
-  // concat<U>(other: Iterable<U>): Iterable<T | U>
-  filter(predicate: (entry: T) => any): IGatsbyIterable<T>
-  forEach(callback: (entry: T) => any): void
+export interface IQueryResult {
+  entries: GatsbyIterable<IGatsbyNode>
+  totalCount: () => Promise<number>
+}
+
+export interface IRunQueryArgs {
+  gqlType: GatsbyGraphQLType
+  queryArgs: {
+    filter: IInputQuery | undefined
+    sort:
+      | {
+          fields: Array<string>
+          order: Array<boolean | "asc" | "desc" | "ASC" | "DESC">
+        }
+      | undefined
+    limit?: number
+    skip?: number
+  }
+  firstOnly: boolean
+  resolvedFields: Record<string, any>
+  nodeTypeNames: Array<string>
+  stats: IGraphQLRunnerStats
 }
 
 export interface IDataStore {
@@ -23,8 +42,9 @@ export interface IDataStore {
   getTypes(): Array<string>
   countNodes(typeName?: string): number
   ready(): Promise<void>
-  iterateNodes(): IGatsbyIterable<IGatsbyNode>
-  iterateNodesByType(type: string): IGatsbyIterable<IGatsbyNode>
+  iterateNodes(): GatsbyIterable<IGatsbyNode>
+  iterateNodesByType(type: string): GatsbyIterable<IGatsbyNode>
+  runQuery(args: IRunQueryArgs): Promise<IQueryResult>
 
   /** @deprecated */
   getNodes(): Array<IGatsbyNode>
