@@ -646,6 +646,7 @@ describe(`NodeModel`, () => {
   describe(`materialization`, () => {
     let resolveBetterTitleMock
     let resolveOtherTitleMock
+    let resolveSlugMock
     beforeEach(async () => {
       const nodes = (() => [
         {
@@ -732,6 +733,7 @@ describe(`NodeModel`, () => {
       )
       resolveBetterTitleMock = jest.fn()
       resolveOtherTitleMock = jest.fn()
+      resolveSlugMock = jest.fn()
       store.dispatch({
         type: `CREATE_TYPES`,
         payload: [
@@ -795,7 +797,10 @@ describe(`NodeModel`, () => {
               },
               slug: {
                 type: `String`,
-                resolve: source => source.id,
+                resolve: source => {
+                  resolveSlugMock()
+                  return source.id
+                },
               },
             },
           }),
@@ -926,6 +931,70 @@ describe(`NodeModel`, () => {
       )
       expect(resolveBetterTitleMock.mock.calls.length).toBe(2)
       expect(resolveOtherTitleMock.mock.calls.length).toBe(2)
+    })
+
+    it(`should not resolve prepared nodes more than once (with mixed interfaces and node types)`, async () => {
+      nodeModel.replaceFiltersCache()
+      await nodeModel.runQuery(
+        {
+          query: { filter: { slug: { eq: `id1` } } },
+          firstOnly: false,
+          type: `Test`,
+        },
+        { path: `/` }
+      )
+      expect(resolveSlugMock.mock.calls.length).toBe(2)
+      expect(resolveBetterTitleMock.mock.calls.length).toBe(0)
+      nodeModel.replaceFiltersCache()
+      await nodeModel.runQuery(
+        {
+          query: { filter: { slug: { eq: `id1` } } },
+          firstOnly: false,
+          type: `TestInterface`,
+        },
+        { path: `/` }
+      )
+      expect(resolveSlugMock.mock.calls.length).toBe(2)
+      expect(resolveBetterTitleMock.mock.calls.length).toBe(0)
+      nodeModel.replaceFiltersCache()
+      await nodeModel.runQuery(
+        {
+          query: {
+            filter: { slug: { eq: `id1` }, betterTitle: { eq: `foo` } },
+          },
+          firstOnly: false,
+          type: `Test`,
+        },
+        { path: `/` }
+      )
+      expect(resolveSlugMock.mock.calls.length).toBe(2)
+      expect(resolveBetterTitleMock.mock.calls.length).toBe(2)
+      nodeModel.replaceFiltersCache()
+      await nodeModel.runQuery(
+        {
+          query: {
+            filter: { slug: { eq: `id1` } },
+          },
+          firstOnly: false,
+          type: `TestInterface`,
+        },
+        { path: `/` }
+      )
+      expect(resolveSlugMock.mock.calls.length).toBe(2)
+      expect(resolveBetterTitleMock.mock.calls.length).toBe(2)
+      nodeModel.replaceFiltersCache()
+      await nodeModel.runQuery(
+        {
+          query: {
+            filter: { slug: { eq: `id1` }, betterTitle: { eq: `foo` } },
+          },
+          firstOnly: true,
+          type: `Test`,
+        },
+        { path: `/` }
+      )
+      expect(resolveSlugMock.mock.calls.length).toBe(2)
+      expect(resolveBetterTitleMock.mock.calls.length).toBe(2)
     })
 
     it(`can filter by resolved fields`, async () => {
@@ -1310,8 +1379,8 @@ describe(`NodeModel`, () => {
         const copiedInlineObject = { ...node.inlineObject }
         nodeModel.trackInlineObjectsInRootNode(copiedInlineObject)
 
-        expect(nodeModel._trackedRootNodes instanceof Set).toBe(true)
-        expect(nodeModel._trackedRootNodes.has(node.id)).toEqual(true)
+        expect(nodeModel._trackedRootNodes instanceof WeakSet).toBe(true)
+        expect(nodeModel._trackedRootNodes.has(node)).toEqual(true)
       })
     })
     describe(`not directly on a node`, () => {
@@ -1354,8 +1423,8 @@ describe(`NodeModel`, () => {
         const copiedInlineObject = { ...node.inlineObject }
         nodeModel.trackInlineObjectsInRootNode(copiedInlineObject)
 
-        expect(nodeModel._trackedRootNodes instanceof Set).toBe(true)
-        expect(nodeModel._trackedRootNodes.has(node.id)).toEqual(true)
+        expect(nodeModel._trackedRootNodes instanceof WeakSet).toBe(true)
+        expect(nodeModel._trackedRootNodes.has(node)).toEqual(true)
       })
     })
   })
