@@ -20,7 +20,7 @@ import {
   getNodesByType,
   getTypes,
 } from "../datastore"
-import { runFastFiltersAndSort } from "../datastore/in-memory/run-fast-filters"
+import { isIterable } from "../datastore/common/iterable"
 
 type TypeOrTypeName = string | GraphQLOutputType
 
@@ -75,7 +75,7 @@ class LocalNodeModel {
     this.createPageDependencyActionCreator = createPageDependency
 
     this._rootNodeMap = new WeakMap()
-    this._trackedRootNodes = new Set()
+    this._trackedRootNodes = new WeakSet()
     this._prepareNodesQueues = {}
     this._prepareNodesPromises = {}
     this._preparedNodesCache = new Map()
@@ -230,7 +230,20 @@ class LocalNodeModel {
    * @returns {Promise<Node[]>}
    */
   async runQuery(args, pageDependencies = {}) {
-    const { query, firstOnly, type, stats, tracer } = args || {}
+    // TODO: show deprecation warning in v4
+    // reporter.warn(
+    //   `nodeModel.runQuery() is deprecated. Use nodeModel.findAll() or nodeModel.findOne() instead`
+    // )
+    if (args.firstOnly) {
+      return this.findOne(args, pageDependencies)
+    }
+    const { skip, limit, ...query } = args.query
+    const result = await this.findAll({ ...args, query }, pageDependencies)
+    return Array.from(result.entries)
+  }
+
+  async _query(args) {
+    const { query, type, stats, tracer } = args || {}
 
     // We don't support querying union types (yet?), because the combined types
     // need not have any fields in common.
@@ -266,7 +279,10 @@ class LocalNodeModel {
       nodeTypeNames
     )
 
-    await this.prepareNodes(gqlType, fields, fieldsToResolve, nodeTypeNames)
+    for (const nodeTypeName of nodeTypeNames) {
+      const gqlNodeType = this.schema.getType(nodeTypeName)
+      await this.prepareNodes(gqlNodeType, fields, fieldsToResolve)
+    }
 
     if (materializationActivity) {
       materializationActivity.end()
@@ -280,9 +296,8 @@ class LocalNodeModel {
       runQueryActivity.start()
     }
 
-    const queryResult = runFastFiltersAndSort({
+    const { entries, totalCount } = await getDataStore().runQuery({
       queryArgs: query,
-      firstOnly,
       gqlSchema: this.schema,
       gqlComposer: this.schemaComposer,
       gqlType,
@@ -296,49 +311,63 @@ class LocalNodeModel {
       runQueryActivity.end()
     }
 
-    let trackInlineObjectsActivity
-    if (tracer) {
-      trackInlineObjectsActivity = reporter.phantomActivity(
-        `trackInlineObjects`,
-        {
-          parentSpan: tracer.getParentActivity().span,
-        }
-      )
-      trackInlineObjectsActivity.start()
+    return {
+      gqlType,
+      entries: entries.map(node => {
+        // With GatsbyIterable it happens lazily as we iterate
+        this.trackInlineObjectsInRootNode(node)
+        return node
+      }),
+      totalCount,
     }
-
-    let result = queryResult
-    if (firstOnly) {
-      if (result?.length > 0) {
-        result = result[0]
-        this.trackInlineObjectsInRootNode(result)
-      } else {
-        result = null
-
-        // Couldn't find matching node.
-        //  This leads to a state where data tracking for this query gets empty.
-        //  It means we will NEVER re-run this query on any data updates
-        //  (even if a new node matching this query is added at some point).
-        //  To workaround this, we have to add a connection tracking to re-run
-        //  the query whenever any node of this type changes.
-        pageDependencies.connectionType = gqlType.name
-      }
-    } else if (result) {
-      result.forEach(node => this.trackInlineObjectsInRootNode(node))
-    }
-
-    if (trackInlineObjectsActivity) {
-      trackInlineObjectsActivity.end()
-    }
-
-    // Tracking connections by default:
-    if (!firstOnly && typeof pageDependencies.connectionType === `undefined`) {
-      pageDependencies.connectionType = gqlType.name
-    }
-    return this.trackPageDependencies(result, pageDependencies)
   }
 
-  prepareNodes(type, queryFields, fieldsToResolve, nodeTypeNames) {
+  async findAll(args, pageDependencies = {}) {
+    // TODO: add this as a public API in v4 (together with deprecating runQuery)
+    const { gqlType, ...result } = await this._query(args, pageDependencies)
+
+    // Tracking connections by default:
+    if (typeof pageDependencies.connectionType === `undefined`) {
+      pageDependencies.connectionType = gqlType.name
+    }
+    this.trackPageDependencies(result.entries, pageDependencies)
+    return result
+  }
+
+  async findOne(args, pageDependencies = {}) {
+    // TODO: add this as a public API in v4 (together with deprecating runQuery)
+    const { query } = args
+    if (query.sort?.fields?.length > 0) {
+      // If we support sorting and return the first node based on sorting
+      // we'll have to always track connection not an individual node
+      reporter.warn(
+        `nodeModel.findOne() does not support sorting. Use nodeModel.findAll({ query: { limit: 1 } }) instead`
+      )
+      // TODO: throw in v4
+      // throw new Error(
+      //   `nodeModel.findOne() does not support sorting. Use nodeModel.findAll({ query: { limit: 1 } }) instead`
+      // )
+    }
+    const { gqlType, entries } = await this._query({
+      ...args,
+      query: { ...query, skip: 0, limit: 1, sort: undefined },
+    })
+    const result = Array.from(entries)
+    const first = result[0] ?? null
+
+    if (!first) {
+      // Couldn't find matching node.
+      //  This leads to a state where data tracking for this query gets empty.
+      //  It means we will NEVER re-run this query on any data updates
+      //  (even if a new node matching this query is added at some point).
+      //  To workaround this, we have to add a connection tracking to re-run
+      //  the query whenever any node of this type changes.
+      pageDependencies.connectionType = gqlType.name
+    }
+    return this.trackPageDependencies(first, pageDependencies)
+  }
+
+  prepareNodes(type, queryFields, fieldsToResolve) {
     const typeName = type.name
     if (!this._prepareNodesQueues[typeName]) {
       this._prepareNodesQueues[typeName] = []
@@ -352,7 +381,7 @@ class LocalNodeModel {
     if (!this._prepareNodesPromises[typeName]) {
       this._prepareNodesPromises[typeName] = new Promise(resolve => {
         process.nextTick(async () => {
-          await this._doResolvePrepareNodesQueue(type, nodeTypeNames)
+          await this._doResolvePrepareNodesQueue(type)
           resolve()
         })
       })
@@ -361,7 +390,7 @@ class LocalNodeModel {
     return this._prepareNodesPromises[typeName]
   }
 
-  async _doResolvePrepareNodesQueue(type, nodeTypeNames) {
+  async _doResolvePrepareNodesQueue(type) {
     const typeName = type.name
     const queue = this._prepareNodesQueues[typeName]
     this._prepareNodesQueues[typeName] = []
@@ -389,7 +418,8 @@ class LocalNodeModel {
     )
 
     if (!_.isEmpty(actualFieldsToResolve)) {
-      await saveResolvedNodes(nodeTypeNames, async node => {
+      const resolvedNodes = new Map()
+      for (const node of getDataStore().iterateNodesByType(typeName)) {
         this.trackInlineObjectsInRootNode(node)
         const resolvedFields = await resolveRecursive(
           this,
@@ -403,8 +433,14 @@ class LocalNodeModel {
         if (!node.__gatsby_resolved) {
           node.__gatsby_resolved = {}
         }
-        return _.merge(node.__gatsby_resolved, resolvedFields)
-      })
+        resolvedNodes.set(
+          node.id,
+          _.merge(node.__gatsby_resolved, resolvedFields)
+        )
+      }
+      if (resolvedNodes.size) {
+        await saveResolvedNodes(typeName, resolvedNodes)
+      }
       this._preparedNodesCache.set(
         typeName,
         _.merge(
@@ -431,7 +467,7 @@ class LocalNodeModel {
    * @param {Node} node Root Node
    */
   trackInlineObjectsInRootNode(node) {
-    if (!this._trackedRootNodes.has(node.id)) {
+    if (!this._trackedRootNodes.has(node)) {
       addRootNodeToInlineObject(
         this._rootNodeMap,
         node,
@@ -439,7 +475,7 @@ class LocalNodeModel {
         true,
         new Set()
       )
-      this._trackedRootNodes.add(node.id)
+      this._trackedRootNodes.add(node)
     }
   }
 
@@ -491,7 +527,7 @@ class LocalNodeModel {
       if (connectionType) {
         this.createPageDependency({ path, connection: connectionType })
       } else {
-        const nodes = Array.isArray(result) ? result : [result]
+        const nodes = isIterable(result) ? result : [result]
         for (const node of nodes) {
           if (node) {
             this.createPageDependency({ path, nodeId: node.id })
@@ -547,6 +583,20 @@ class ContextualNodeModel {
 
   runQuery(args, pageDependencies) {
     return this.nodeModel.runQuery(
+      args,
+      this._getFullDependencies(pageDependencies)
+    )
+  }
+
+  findOne(args, pageDependencies) {
+    return this.nodeModel.findOne(
+      args,
+      this._getFullDependencies(pageDependencies)
+    )
+  }
+
+  findAll(args, pageDependencies) {
+    return this.nodeModel.findAll(
       args,
       this._getFullDependencies(pageDependencies)
     )
@@ -868,23 +918,14 @@ const addRootNodeToInlineObject = (
   }
 }
 
-const saveResolvedNodes = async (nodeTypeNames, resolver) => {
-  for (const typeName of nodeTypeNames) {
-    const resolvedNodes = new Map()
-    for (const node of getDataStore().iterateNodesByType(typeName)) {
-      const resolved = await resolver(node)
-      resolvedNodes.set(node.id, resolved)
-    }
-    if (!resolvedNodes.size) continue
-
-    store.dispatch({
-      type: `SET_RESOLVED_NODES`,
-      payload: {
-        key: typeName,
-        nodes: resolvedNodes,
-      },
-    })
-  }
+const saveResolvedNodes = async (typeName, resolvedNodes) => {
+  store.dispatch({
+    type: `SET_RESOLVED_NODES`,
+    payload: {
+      key: typeName,
+      nodes: resolvedNodes,
+    },
+  })
 }
 
 const deepObjectDifference = (from, to) => {
