@@ -26,7 +26,7 @@ itWhenLMDB(`worker can access node created in main process`, async () => {
   const testNodeId = `shared-node`
 
   expect(getDataStore().getNode(testNodeId)).toBeFalsy()
-  expect(await worker.getNodeFromWorker(testNodeId)).toBeFalsy()
+  expect(await worker.single.getNodeFromWorker(testNodeId)).toBeFalsy()
 
   const node = {
     id: testNodeId,
@@ -39,7 +39,9 @@ itWhenLMDB(`worker can access node created in main process`, async () => {
   await getDataStore().ready()
 
   const nodeStoredInMainProcess = getDataStore().getNode(testNodeId)
-  const nodeStoredInWorkerProcess = await worker.getNodeFromWorker(testNodeId)
+  const nodeStoredInWorkerProcess = await worker.single.getNodeFromWorker(
+    testNodeId
+  )
 
   expect(nodeStoredInWorkerProcess).toMatchInlineSnapshot(`
     Object {
