@@ -49,15 +49,15 @@ describeWhenLMDB(`worker (schema)`, () => {
 
     const siteDirectory = path.join(__dirname, `fixtures`, `sample-site`)
     await loadConfigAndPlugins({ siteDirectory })
-    await worker.loadConfigAndPlugins({ siteDirectory })
+    await Promise.all(worker.all.loadConfigAndPlugins({ siteDirectory }))
     await sourceNodesAndRemoveStaleNodes({ webhookBody: {} })
     await getDataStore().ready()
 
     await build({ parentSpan: {} })
     saveStateForWorkers([`inferenceMetadata`])
-    await worker.buildSchema()
+    await Promise.all(worker.all.buildSchema())
 
-    stateFromWorker = await worker.getState()
+    stateFromWorker = await worker.single.getState()
   })
 
   afterAll(() => {
@@ -130,7 +130,7 @@ describeWhenLMDB(`worker (schema)`, () => {
 
   it(`should have resolverField from createResolvers`, async () => {
     // @ts-ignore - it exists
-    const { data } = await worker?.getRunQueryResult(`
+    const { data } = await worker?.single.getRunQueryResult(`
     {
       one: nodeTypeOne {
         number
@@ -151,7 +151,7 @@ describeWhenLMDB(`worker (schema)`, () => {
 
   it(`should have fieldsOnGraphQL from setFieldsOnGraphQLNodeType`, async () => {
     // @ts-ignore - it exists
-    const { data } = await worker?.getRunQueryResult(`
+    const { data } = await worker?.single.getRunQueryResult(`
     {
       four: nodeTypeOne {
         fieldsOnGraphQL
